debian: add packaging
Add initial debian packaging and a systemd service that starts
osmo_dia2gsup. Given that dependencies listed in rebar.lock are not
packaged in debian 11, this just uses rebar3 to download and compile
the dependencies during the build.

Resulting erlang files are installed to the following path, where they
become available in the erlang shell (erl):
  /lib/erlang/lib/{name_of_component}/ebin/

Related: SYS#6006
Change-Id: Ic55bca1260f992d0ed8583e7c7fe948ba0160d4d